Moved commandFactory to model
