+++ src/dependencyInjection/index.ts
@@ -1,4 +1,4 @@
 export * from "./types"
 export * from "./inversify.config"
-export * from "./basicPluginDependencies"
-export * from "./pluginDependencies"
+export * from "../model/services/basicPluginDependencies"
+export * from "../model/services/pluginDependencies"
