+++ src/dependencyInjection/inversify.config.ts
@@ -7,8 +7,8 @@ import { RepositoryImpl } from "../model/data/repositoryImpl"
 import { CommandFactory, CommandFactoryImpl } from "../model/commandFramework"
 import { SpotifyAPI } from "../model/services/spotify/spotifyApi"
 import { SpotifyAPIImpl } from "../model/services/spotify/spotifyApiImpl"
-import { PluginDependencies } from "./pluginDependencies"
-import { BasicPluginDependencies } from "./basicPluginDependencies"
+import { PluginDependencies } from "../model/services/pluginDependencies"
+import { BasicPluginDependencies } from "../model/services/basicPluginDependencies"
 import { DiscordWrapper } from "../controller/discordWrapper"
 import { DiscordWrapperImpl } from "../controller/discordWrapperImpl"
 import { HttpController } from "../controller/httpController"
